# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Confidence": {
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "Confidence",
        "type": "string"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Código estable de error, apto para lógica de cliente.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "description": "Información adicional, opcional.",
            "title": "Details"
          },
          "message": {
            "description": "Mensaje presentable al usuario.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FindingCategory": {
        "enum": [
          "security",
          "seo",
          "performance",
          "accessibility",
          "best_practices"
        ],
        "title": "FindingCategory",
        "type": "string"
      },
...
      "FindingRead": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FindingCategory"
          },
          "client_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Explanation"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cwe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwe"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "occurrences": {
            "title": "Occurrences",
            "type": "integer"
          },
          "owasp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owasp"
          },
          "parameter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameter"
          },
          "references": {
            "items": {},
            "title": "References",
            "type": "array"
          },
          "remediation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "scan_id": {
            "format": "uuid",
            "title": "Scan Id",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "source": {
            "$ref": "#/components/schemas/FindingSource"
          },
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "scan_id",
          "source",
          "category",
          "rule_id",
          "title",
          "severity",
          "confidence",
          "url",
          "parameter",
          "evidence",
          "description",
          "impact",
          "remediation",
          "client_explanation",
          "cwe",
          "owasp",
          "references",
          "occurrences",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "FindingRead",
        "type": "object"
      },
      "FindingSource": {
        "enum": [
          "zap",
          "seo",
          "crawler",
          "pagespeed",
          "search_console"
        ],
        "title": "FindingSource",
        "type": "string"
      },
      "FindingStatus": {
        "enum": [
          "open",
          "fixed",
          "accepted",
          "false_positive"
        ],
        "title": "FindingStatus",
        "type": "string"
      },
      "FindingStatusUpdate": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "FindingStatusUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_FindingRead_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FindingRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[FindingRead]",
        "type": "object"
      },
...
      "Severity": {
        "enum": [
          "critical",
          "high",
          "medium",
          "low",
          "info"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API interna de **Softree Audit**: auditoría de seguridad, SEO y rendimiento de\nsitios web propios, entregados por Softree o autorizados explícitamente por el\ncliente.\n\nAutenticación mediante `Authorization: Bearer <access_token>` obtenido en\n`POST /api/v1/auth/login`.",
    "title": "Softree Audit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/findings": {
      "get": {
        "operationId": "list_findings_api_v1_findings_get",
        "parameters": [
          {
            "in": "query",
            "name": "scan_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scan Id"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FindingRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Listar hallazgos",
        "tags": [
          "findings"
        ]
      }
    },
    "/api/v1/findings/{finding_id}": {
      "patch": {
        "description": "Permite marcar un hallazgo como corregido, aceptado o falso positivo.",
        "operationId": "update_finding_api_v1_findings__finding_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El hallazgo no existe"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Cambiar el estado de un hallazgo",
        "tags": [
          "findings"
        ]
      }
    },
...
    "/api/v1/scans/{scan_id}/findings": {
      "get": {
        "operationId": "scan_findings_api_v1_scans__scan_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scan Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FindingRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "La auditoría no existe"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Hallazgos de una auditoría",
        "tags": [
          "scans"
        ]
      }
    },
...
  }
}